Donation component state management changes
