+++ components/donation/DonationPage.tsx
@@ -1,4 +1,4 @@
-import React, { useReducer } from "react";
+import React from "react";
 
 import makeStyles from "@material-ui/core/styles/makeStyles";
 
@@ -8,8 +8,6 @@ import { Nonprofit, DropdownProps } from "utils/types";
 import DonationPageLayout from "./DonationPageLayout";
 import DonationPageFormHeader from "./DonationPageFormHeader";
 import DonationPageFormBody from "./DonationPageFormBody";
-
-import reducer, { initialState, DonationPageStateDispatch } from "./reducer";
 
 const useStyles = makeStyles({
   container: {
@@ -27,21 +25,19 @@ const DonationPage: React.FC<Props & DropdownProps> = ({
   ...dropdownProps
 }) => {
   const { container } = useStyles();
-  const [state, dispatch] = useReducer(reducer, initialState);
 
+  // Setting key will tell React to re-mount the DonationPageFormBody component when the selected Nonprofit in the dropdown changes
   return (
-    <DonationPageStateDispatch.Provider value={dispatch}>
-      <DonationPageLayout {...dropdownProps}>
-        <ContainerWithShadow className={container}>
-          <DonationPageFormHeader headline={nonprofit.headline} />
-          <DonationPageFormBody
-            state={state}
-            description={nonprofit.about}
-            selectedNonprofitId={dropdownProps.selectedValue}
-          />
-        </ContainerWithShadow>
-      </DonationPageLayout>
-    </DonationPageStateDispatch.Provider>
+    <DonationPageLayout {...dropdownProps}>
+      <ContainerWithShadow className={container}>
+        <DonationPageFormHeader headline={nonprofit.headline} />
+        <DonationPageFormBody
+          description={nonprofit.about}
+          selectedNonprofitId={dropdownProps.selectedValue}
+          key={dropdownProps.selectedValue}
+        />
+      </ContainerWithShadow>
+    </DonationPageLayout>
   );
 };
